src/requirements.txt
flask==2.1.2
google-cloud-bigquery
pandas
db-dtypes
fastapi
uvicorn
pydantic